Add splash screen, polish UI and text style
